Let the server know WHO fires in order to apply the correct ink color to the drop.
- The color for each player is assigned and replicated when spawning a
player
- The server creates the projectile and assigns that player id as the
owner of that projectile
- When it comes to assign the owner to the ink, the same one as the
projectile is used. At this point, the color is applied by obtaining it
first from the game state according to the player id.
- At the moment only works for the server

+++ Source/ColouringBook/ColouringBookInkDrop.h
@@ -23,11 +23,11 @@ public:
 	AColouringBookInkDrop();
 
 	// Sets the owner
-	void SetOwnerID(uint8 _ownerID);
+	void SetOwnerID(uint32 _ownerID);
 
 private:
 	/* The number ID of the actor that spawned this actor */
-	uint8 ownerID;
+	uint32 ownerID;
 
 	/* The dynamic material instance of the ink drop */
 	class UMaterialInstanceDynamic* dynamicMaterial;
