+++ Source/ColouringBook/ColouringBookProjectile.h
@@ -35,14 +35,14 @@ public:
 	UFUNCTION()
 	void OnHit(UPrimitiveComponent* HitComp, AActor* OtherActor, UPrimitiveComponent* OtherComp, FVector NormalImpulse, const FHitResult& Hit);
 
-	void SetOwnerID(uint8 _ownerID) { ownerID = _ownerID; }
+	void SetOwnerID(uint32 _ownerID) { ownerID = _ownerID; }
 
 private:
 	/* Spawns the ink drops from this bullet */
 	void SpawnInkDrops();
 
 	/* Sets the number ID of the actor that spawned this actor */
-	uint8 ownerID;
+	uint32 ownerID;
 
 public:
 	/** Returns ProjectileMesh subobject **/
